complete GetArticles api

+++ core/internal/types/types.go
@@ -53,5 +53,5 @@ type LoginResponseData struct {
 	Avatar    string `json:"avatar"`
 	UserName  string `json:"username"`
 	Token     string `json:"token"`
-	ExpiredAt int    `json:"exp"`
+	ExpiredAt int64    `json:"exp"`
 }
